+++ include/cpplib/math/number-theory/prime-number/wheel.hpp
@@ -9,10 +9,10 @@
  * Time Complexity: O(sqrt(x)).
  * Space Complexity: O(log(x)).
  */
-vector<pair<uint64_t, uint64_t> > factorization(uint64_t x){
+vector<pair<uint, uint> > factorization(uint x){
     assert(x > 0);
-    vector<pair<uint64_t, uint64_t> > factors;
-    for(uint64_t p : {2, 3, 5}){
+    vector<pair<uint, uint> > factors;
+    for(uint p : {2, 3, 5}){
         while(x%p == 0){
             if(factors.empty() or factors.back().first != p)
                 factors.emplace_back(p, 1);
@@ -21,8 +21,8 @@ vector<pair<uint64_t, uint64_t> > factorization(uint64_t x){
             x /= p;
         }
     }
-    vector<uint64_t> increments = {4, 2, 4, 2, 4, 6, 2, 6};
-    for(uint64_t p = 7, i = 0; p*p <= x; p += increments[i++], i %= 8){
+    vector<uint> increments = {4, 2, 4, 2, 4, 6, 2, 6};
+    for(uint p = 7, i = 0; p*p <= x; p += increments[i++], i %= 8){
         while(x%p == 0){
             if(factors.empty() or factors.back().first != p)
                 factors.emplace_back(p, 1);
